refactor(specification): deduplicate deference mechanism
Refs https://github.com/swagger-api/oss-planning/issues/133
