+++ apidom/packages/apidom-ast/src/index.ts
@@ -61,8 +61,3 @@ export { default as Error } from './Error';
 export { default as ParseResult } from './ParseResult';
 // AST traversal related exports
 export { getVisitFn, BREAK, visit, isNode } from './visitor';
-// YAML CST/AST transformers related exports
-export {
-  transform as transformTreeSitterYamlCST,
-  keyMap as treeSitterYamlKeyMap,
-} from './yaml/transformers/tree-sitter-yaml';
