+++ apidom/packages/apidom-ns-asyncapi-2-0/src/refractor/specification.ts
@@ -1,6 +1,3 @@
-import { mapObjIndexed, has, path, defaultTo, propSatisfies } from 'ramda';
-import { trimCharsStart, isPlainObject, isString } from 'ramda-adjunct';
-
 import FallbackVisitor from './visitors/FallbackVisitor';
 import SpecificationExtensionVisitor from './visitors/SpecificationExtensionVisitor';
 import AsyncApi2_0Visitor from './visitors/async-api-2-0';
@@ -218,26 +215,6 @@ const specification = {
       },
     },
   },
-};
-
-export const dereference = (
-  object: Record<string, any>,
-  root?: Record<string, any>,
-): Record<string, any> => {
-  const rootObject = defaultTo(object, root);
-
-  return mapObjIndexed((val) => {
-    if (isPlainObject(val) && has('$ref', val) && propSatisfies(isString, '$ref', val)) {
-      const $ref = path(['$ref'], val);
-      // @ts-ignore
-      const pointer = trimCharsStart('#/', $ref);
-      return path(pointer.split('/'), rootObject);
-    }
-    if (isPlainObject(val)) {
-      return dereference(val, rootObject);
-    }
-    return val;
-  }, object);
 };
 
 export default specification;
