+++ apidom/packages/apidom-parser-adapter-json/src/parser/specification.ts
@@ -1,6 +1,3 @@
-import { defaultTo, has, mapObjIndexed, path, propSatisfies } from 'ramda';
-import { isPlainObject, isString, trimCharsStart } from 'ramda-adjunct';
-
 import DocumentVisitor from './visitors/DocumentVisitor';
 import ErrorVisitor from './visitors/ErrorVisitor';
 import { ValueVisitor, ObjectVisitor, ArrayVisitor } from './visitors/generics';
@@ -24,26 +21,6 @@ const specification = {
       extension: ValueVisitor,
     },
   },
-};
-
-export const dereference = (
-  object: Record<string, any>,
-  root?: Record<string, any>,
-): Record<string, any> => {
-  const rootObject = defaultTo(object, root);
-
-  return mapObjIndexed((val) => {
-    if (isPlainObject(val) && has('$ref', val) && propSatisfies(isString, '$ref', val)) {
-      const $ref = path(['$ref'], val);
-      // @ts-ignore
-      const pointer = trimCharsStart('#/', $ref);
-      return path(pointer.split('/'), rootObject);
-    }
-    if (isPlainObject(val)) {
-      return dereference(val, rootObject);
-    }
-    return val;
-  }, object);
 };
 
 export default specification;
