+++ apidom/packages/apidom/src/index.ts
@@ -48,6 +48,7 @@ export { default as createPredicate } from './predicates/helpers';
 export { filter, reject, find, findAtOffset, some, traverse } from './traversal';
 export { visit, BREAK, getNodeType, keyMapDefault as keyMap } from './traversal/visitor';
 export { transclude, default as Transcluder } from './transcluder';
+export { dereference } from './util';
 
 export const createNamespace = (namespacePlugin?: NamespacePlugin): ApiDOMNamespace => {
   const namespace = new ApiDOMNamespace();
